+++ exe/Main.hs
@@ -84,6 +84,8 @@ data KnownType
   | SPDX
   | LicenseId
   | LicenseExceptionId
+  | Scope
+  | ModuleRenaming
   deriving (Bounded, Enum, Eq, Ord, Read, Show)
 
 
@@ -463,6 +465,8 @@ printType PrintTypeOptions { .. } = do
           SPDX -> Dhall.expected spdxLicense
           LicenseId -> Dhall.expected spdxLicenseId
           LicenseExceptionId -> Dhall.expected spdxLicenseExceptionId
+          Scope -> Dhall.expected executableScope
+          ModuleRenaming -> Dhall.expected moduleRenaming
       )
 
     makeLetOrImport t val reduced =