+++ Changelog.md
@@ -9,6 +9,8 @@
 
 * Use `dhall` version 1.23.0.
 
+* Educated `--print-type` about `Scope` and `ModuleRenaming`.
+
 * Fix `cabal-to-dhall` output for unknown licenses using `cabal` spec
   version below `2.0`. Now it uses values of type `License.Unknown Text` 
   for them.